// src/app/core/api/model/variableSetSummaryDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface VariableSetSummaryDto { 
    /**
     * Variable set name
     */
    name: string;
    scope: VariableSetSummaryDto.ScopeEnum;
    type: VariableSetSummaryDto.TypeEnum;
    resourceVersion?: string;
    /**
     * Key names present in this variable set
     */
    keys: Array<string>;
    /**
     * Variable values — only present for type=plain
     */
    data?: { [key: string]: string; };
}
export namespace VariableSetSummaryDto {
    export const ScopeEnum = {
        App: 'app',
        System: 'system',
        Shared: 'shared'
    } as const;
    export type ScopeEnum = typeof ScopeEnum[keyof typeof ScopeEnum];
    export const TypeEnum = {
        Plain: 'plain',
        Sensitive: 'sensitive',
        All: 'all'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
}