// src/app/core/api/model/appVariablesResponseDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AppVariablesResponseDto { 
    /**
     * Internal resource name
     */
    name: string;
    type: AppVariablesResponseDto.TypeEnum;
    scope: AppVariablesResponseDto.ScopeEnum;
    /**
     * Variable values (only returned for plain type)
     */
    data?: { [key: string]: string; };
    /**
     * Variable keys (returned for sensitive type — values are never exposed)
     */
    keys?: Array<string>;
    resourceVersion?: string;
}
export namespace AppVariablesResponseDto {
    export const TypeEnum = {
        Plain: 'plain',
        Sensitive: 'sensitive',
        All: 'all'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
    export const ScopeEnum = {
        App: 'app',
        System: 'system',
        Shared: 'shared'
    } as const;
    export type ScopeEnum = typeof ScopeEnum[keyof typeof ScopeEnum];
}